+++ src/import_dats/test_dat_headered_skipped_header.rs
@@ -1,4 +1,3 @@
-use super::super::config::*;
 use super::super::database::*;
 use super::*;
 use std::path::PathBuf;