+++ src/import_roms/test_chd_multiple_tracks_without_cue_should_fail.rs
@@ -48,6 +48,7 @@ async fn test() {
         &romfile_path,
         &HashAlgorithm::Crc,
         true,
+        false,
     )
     .await
     .unwrap();