+++ src/import_roms/test_chd_single_track.rs
@@ -50,6 +50,7 @@ async fn test() {
         &romfile_path,
         &HashAlgorithm::Crc,
         true,
+        false,
     )
     .await
     .unwrap();