+++ src/import_roms/test_cia.rs
@@ -59,6 +59,7 @@ async fn test() {
         &romfile_path,
         &HashAlgorithm::Crc,
         true,
+        false,
     )
     .await
     .unwrap();