+++ src/import_roms/test_original_headered.rs
@@ -51,6 +51,7 @@ async fn test() {
         &romfile_path,
         &HashAlgorithm::Crc,
         true,
+        false,
     )
     .await
     .unwrap();