+++ src/import_roms/test_sevenzip_multiple_files_partial_game.rs
@@ -51,6 +51,7 @@ async fn test() {
         romfile_path.extension().unwrap().to_str().unwrap(),
         &HashAlgorithm::Crc,
         true,
+        false,
     )
     .await
     .unwrap();