+++ src/import_roms/test_sevenzip_single_file_headered.rs
@@ -52,6 +52,7 @@ async fn test() {
         romfile_path.extension().unwrap().to_str().unwrap(),
         &HashAlgorithm::Crc,
         true,
+        false,
     )
     .await
     .unwrap();