+++ src/import_roms/test_zso.rs
@@ -58,6 +58,7 @@ async fn test() {
         &romfile_path,
         &HashAlgorithm::Crc,
         true,
+        false,
     )
     .await
     .unwrap();